Fix data loading in simulation (on macOS).
Data loading was broken when I played with this notebook. The filepath
description was not working - I updated things to work with the pathlib
module, which I understand should be platform agnostic. Hopefully this
is the case, and updating to work on my platform (macOS) will not break
things for those developing on windows.